Пятое задание выполнено и проверено.

+++ src/main/java/org/itstep/task02/EmployeeAndTax.java
@@ -6,14 +6,14 @@ import org.itstep.task01.Employee;
 import java.util.Locale;
 
 public class EmployeeAndTax extends Employee {
-    private int tax;
+    private double tax;
 
-    public EmployeeAndTax(String fullName, String paymentType, int payment, int tax) {
+    public EmployeeAndTax(String fullName, String paymentType, double payment, double tax) {
         super(fullName, paymentType, payment);
         this.tax = tax;
     }
 
-    public EmployeeAndTax(String fullName, String paymentType, int payment) {
+    public EmployeeAndTax(String fullName, String paymentType, double payment) {
         super(fullName, paymentType, payment);
         if (paymentType.equals(PaymentTypes.HOURLY)) {
             tax = 20;
@@ -22,11 +22,11 @@ public class EmployeeAndTax extends Employee {
         }
     }
 
-    public int getTax() {
+    public double getTax() {
         return tax;
     }
 
-    public void setTax(int tax) {
+    public void setTax(double tax) {
         this.tax = tax;
     }
 
@@ -40,13 +40,13 @@ public class EmployeeAndTax extends Employee {
     }
 
     public double getPaymentWithTax(double payment) {
-        return payment - payment * (tax / 100.0);
+        return payment - payment * (tax / 100);
     }
 
     @Override
     public String toString() {
         Locale.setDefault(Locale.US);
-        return String.format("%-8s | %8s%s | %-11s | %.1f", getFullName(), getTax(),
+        return String.format("%-8s | %8s%s | %-11s | %.1f", getFullName(), (int) getTax(),
                 "%", getPayment(), getPaymentWithTax(getPayment()));
     }
 }
