+++ src/main/java/org/itstep/task04/EmployeeAndTugrik.java
@@ -5,21 +5,21 @@ import org.itstep.task03.EmployeeAndChild;
 import java.util.Locale;
 
 public class EmployeeAndTugrik extends EmployeeAndChild {
-    private int tugrikExchange;
+    private double tugrikExchange;
 
-    public EmployeeAndTugrik(String fullName, String paymentType, int payment, boolean hasChild) {
+    public EmployeeAndTugrik(String fullName, String paymentType, double payment, boolean hasChild) {
         super(fullName, paymentType, payment, hasChild);
     }
 
-    public EmployeeAndTugrik(String fullName, String paymentType, int payment) {
+    public EmployeeAndTugrik(String fullName, String paymentType, double payment) {
         super(fullName, paymentType, payment);
     }
 
-    public int getTugrikExchange() {
+    public double getTugrikExchange() {
         return tugrikExchange;
     }
 
-    public void setTugrikExchange(int tugrikExchange) {
+    public void setTugrikExchange(double tugrikExchange) {
         this.tugrikExchange = tugrikExchange;
     }
 
@@ -27,10 +27,10 @@ public class EmployeeAndTugrik extends EmployeeAndChild {
     public String toString() {
         Locale.setDefault(Locale.US);
         if (getTugrikExchange() == 0) {
-            return String.format("%-8s | %8s%s | %-11s | %.1f", getFullName(), getTax(),
+            return String.format("%-8s | %8s%s | %-11s | %.1f", getFullName(), (int) getTax(),
                     "%", getPayment(), getPaymentWithTax(getPayment()));
         } else {
-            return String.format("%-8s | %8s%s | %-11s | %.1f / %.1f", getFullName(), getTax(),
+            return String.format("%-8s | %8s%s | %-11s | %.1f / %.1f", getFullName(), (int) getTax(),
                     "%", getPayment(), getPaymentWithTax(getPayment()) / 2,
                     getPaymentWithTax(getPayment()) / 2 / getTugrikExchange());
         }
